add reward dependency
