+++ x/campaign/keeper/keeper.go
@@ -9,17 +9,41 @@ import (
 	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/tendermint/spn/x/campaign/types"
+	launchtypes "github.com/tendermint/spn/x/launch/types"
+	rewardtypes "github.com/tendermint/spn/x/reward/types"
 )
+
+type LaunchKeeper interface {
+	GetChain(ctx sdk.Context, launchID uint64) (val launchtypes.Chain, found bool)
+	CreateNewChain(
+		ctx sdk.Context,
+		coordinatorID uint64,
+		genesisChainID,
+		sourceURL,
+		sourceHash,
+		genesisURL,
+		genesisHash string,
+		hasCampaign bool,
+		campaignID uint64,
+		isMainnet bool,
+		metadata []byte,
+	) (uint64, error)
+}
+
+type RewardKeeper interface {
+	GetRewardPool(ctx sdk.Context, launchID uint64) (val rewardtypes.RewardPool, found bool)
+}
 
 type (
 	Keeper struct {
 		cdc           codec.BinaryCodec
 		storeKey      sdk.StoreKey
 		memKey        sdk.StoreKey
-		launchKeeper  types.LaunchKeeper
+		launchKeeper  LaunchKeeper
 		bankKeeper    types.BankKeeper
 		distrKeeper   types.DistributionKeeper
 		profileKeeper types.ProfileKeeper
+		rewardKeeper  RewardKeeper
 		paramSpace    paramtypes.Subspace
 	}
 )
@@ -29,10 +53,11 @@ func NewKeeper(
 	storeKey,
 	memKey sdk.StoreKey,
 	paramSpace paramtypes.Subspace,
-	launchKeeper types.LaunchKeeper,
+	launchKeeper LaunchKeeper,
 	bankKeeper types.BankKeeper,
 	distrKeeper types.DistributionKeeper,
 	profileKeeper types.ProfileKeeper,
+	rewardKeeper RewardKeeper,
 ) *Keeper {
 	// set KeyTable if it has not already been set
 	if !paramSpace.HasKeyTable() {
@@ -48,6 +73,7 @@ func NewKeeper(
 		bankKeeper:    bankKeeper,
 		distrKeeper:   distrKeeper,
 		profileKeeper: profileKeeper,
+		rewardKeeper:  rewardKeeper,
 	}
 }
