+++ x/campaign/types/expected_keepers.go
@@ -2,26 +2,10 @@ package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	profiletypes "github.com/tendermint/spn/x/profile/types"
 )
-
-type LaunchKeeper interface {
-	CreateNewChain(
-		ctx sdk.Context,
-		coordinatorID uint64,
-		genesisChainID,
-		sourceURL,
-		sourceHash,
-		genesisURL,
-		genesisHash string,
-		hasCampaign bool,
-		campaignID uint64,
-		isMainnet bool,
-		metadata []byte,
-	) (uint64, error)
-}
 
 type BankKeeper interface {
 	GetSupply(ctx sdk.Context, denom string) sdk.Coin
